% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internals.R
\name{binder_installR}
\alias{binder_installR}
\title{Creates install.R file}
\usage{
binder_installR(directory = utils::globalVariables(c(".")))
}
\arguments{
\item{directory}{To build files in. Defaults to current directory.}
}
\description{
Creates install.R file
}
