% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internals.R
\name{binder_runtime}
\alias{binder_runtime}
\title{binder_runtime}
\usage{
binder_runtime(directory)
}
\arguments{
\item{directory}{the directory to write the binder docker file.}
}
\value{
nothing, writes a file.
}
\description{
Create the binder runtime
}
\details{
Writes the binder R runtime.txt to the user's disk current working directory.
}
