% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_binder.R
\name{build_binder}
\alias{build_binder}
\title{Build binder files}
\usage{
build_binder(directory = ".")
}
\arguments{
\item{directory}{To build files in. Defaults to current directory.}
}
\value{
message confirming files created and suggesting next steps
}
\description{
Builds the binder dockerfile and the install.R script.
}
\examples{
build_binder()
}
