% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internals.R
\name{CRAN_package}
\alias{CRAN_package}
\title{Find out if package is on CRAN and get version}
\usage{
CRAN_package(package_name)
}
\arguments{
\item{package_name}{name of package to test}
}
\description{
Find out if package is on CRAN and get version
}
