% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internals.R
\name{find_doc_libs}
\alias{find_doc_libs}
\title{Scan text to find packages}
\usage{
find_doc_libs(doc)
}
\arguments{
\item{doc}{text of file to scan}
}
\description{
Scan text to find packages
}
\note{
from milesmcbain/deplearning
}
