% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internals.R
\name{github_package}
\alias{github_package}
\title{Find out if package is on github and get repo and commit SHA}
\usage{
github_package(package_name)
}
\arguments{
\item{package_name}{name of package to test}
}
\description{
Find out if package is on github and get repo and commit SHA
}
