% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launch_binder.R
\name{launch_binder}
\alias{launch_binder}
\title{Launch binder for project at mybinder.org}
\usage{
launch_binder(github_username, repo_name, branch_name)
}
\arguments{
\item{github_username}{For project repo}

\item{repo_name}{Repository name}

\item{branch_name}{Branch name, eg. 'master'}
}
\value{
opens URL for binder
}
\description{
launches binder
}
